api/requirements.txt
Flask==3.0.0
cachetools==5.3.2
python-dotenv==1.0.0
web3==6.11.3
pytest==7.4.3
pytest-mock==3.12.0